+++ examples/data-sources/meraki_appliance_static_routes/data-source.tf
@@ -1,0 +1,3 @@
+data "meraki_appliance_static_routes" "example" {
+  network_id = "L_123456"
+}