+++ examples/resources/meraki_appliance_static_route/import.sh
@@ -1,0 +1,1 @@
+terraform import meraki_appliance_static_route.example "<network_id>,<id>"